requirements.txt
agentscope==1.0.18
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aioitertools==0.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.92.0
anyio==4.13.0
attrs==26.1.0
bidict==0.23.1
cached-property==2.0.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
dashscope==1.25.16
distro==1.9.0
docstring_parser==0.17.0
filetype==1.2.0
frozenlist==1.8.0
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
jiter==0.13.0
json5==0.14.0
json_repair==0.58.7
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
loguru==0.7.3
mcp==1.27.0
multidict==6.7.1
numpy==2.4.4
openai==2.31.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
propcache==0.4.1
protobuf==6.33.6
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
python-datauri==3.0.2
python-dotenv==1.2.2
python-engineio==4.13.1
python-frontmatter==1.1.0
python-multipart==0.0.24
python-socketio==5.16.1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rpds-py==0.30.0
shortuuid==1.0.13
simple-websocket==1.1.0
sniffio==1.3.1
sounddevice==0.5.5
SQLAlchemy==2.0.49
sse-starlette==3.3.4
starlette==1.0.0
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
websocket-client==1.9.0
wsproto==1.3.2
yarl==1.23.0
zipp==3.23.0
